# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Case Management API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: cases
    description: Legal cases
  - name: documents
    description: Case documents
  - name: deadlines
    description: Case deadlines
  - name: tasks
    description: Case tasks
  - name: interactions
    description: Client interactions
  - name: dashboard
    description: Dashboard summary data

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  # ── Dashboard ──────────────────────────────────────────────────────────────
  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get dashboard summary stats
      responses:
        "200":
          description: Summary stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"

  # ── Cases ──────────────────────────────────────────────────────────────────
  /cases:
    get:
      operationId: listCases
      tags: [cases]
      summary: List all cases
      responses:
        "200":
          description: List of cases
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Case"
    post:
      operationId: createCase
      tags: [cases]
      summary: Create a new case
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CaseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Case"

  /cases/{id}:
    get:
      operationId: getCase
      tags: [cases]
      summary: Get a case by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: The case
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Case"
        "404":
          description: Not found
    patch:
      operationId: updateCase
      tags: [cases]
      summary: Update a case
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CaseUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Case"
    delete:
      operationId: deleteCase
      tags: [cases]
      summary: Delete a case
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Documents ─────────────────────────────────────────────────────────────
  /documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List documents, optionally filtered by case
      parameters:
        - name: caseId
          in: query
          required: false
          schema:
            type: integer
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
    post:
      operationId: createDocument
      tags: [documents]
      summary: Create a document record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"

  /documents/{id}:
    get:
      operationId: getDocument
      tags: [documents]
      summary: Get a document by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: The document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "404":
          description: Not found
    patch:
      operationId: updateDocument
      tags: [documents]
      summary: Update a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Delete a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Deadlines ──────────────────────────────────────────────────────────────
  /deadlines:
    get:
      operationId: listDeadlines
      tags: [deadlines]
      summary: List deadlines, optionally filtered by case
      parameters:
        - name: caseId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of deadlines
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Deadline"
    post:
      operationId: createDeadline
      tags: [deadlines]
      summary: Create a deadline
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeadlineInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deadline"

  /deadlines/{id}:
    get:
      operationId: getDeadline
      tags: [deadlines]
      summary: Get a deadline by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: The deadline
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deadline"
        "404":
          description: Not found
    patch:
      operationId: updateDeadline
      tags: [deadlines]
      summary: Update a deadline
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeadlineUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deadline"
    delete:
      operationId: deleteDeadline
      tags: [deadlines]
      summary: Delete a deadline
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Tasks ─────────────────────────────────────────────────────────────────
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List tasks, optionally filtered by case
      parameters:
        - name: caseId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"

  /tasks/{id}:
    get:
      operationId: getTask
      tags: [tasks]
      summary: Get a task by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: The task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: Not found
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Interactions ───────────────────────────────────────────────────────────
  /interactions:
    get:
      operationId: listInteractions
      tags: [interactions]
      summary: List interactions, optionally filtered by case
      parameters:
        - name: caseId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of interactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Interaction"
    post:
      operationId: createInteraction
      tags: [interactions]
      summary: Create an interaction record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InteractionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Interaction"

  /interactions/{id}:
    get:
      operationId: getInteraction
      tags: [interactions]
      summary: Get an interaction by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: The interaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Interaction"
        "404":
          description: Not found
    patch:
      operationId: updateInteraction
      tags: [interactions]
      summary: Update an interaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InteractionUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Interaction"
    delete:
      operationId: deleteInteraction
      tags: [interactions]
      summary: Delete an interaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    DashboardSummary:
      type: object
      required:
        - totalCases
        - activeCases
        - pendingCases
        - upcomingDeadlines
        - highPriorityDeadlines
        - openTasks
        - inProgressTasks
        - totalDocuments
        - totalInteractions
        - recentActivity
      properties:
        totalCases:
          type: integer
        activeCases:
          type: integer
        pendingCases:
          type: integer
        upcomingDeadlines:
          type: integer
        highPriorityDeadlines:
          type: integer
        openTasks:
          type: integer
        inProgressTasks:
          type: integer
        totalDocuments:
          type: integer
        totalInteractions:
          type: integer
        recentActivity:
          type: array
          items:
            $ref: "#/components/schemas/ActivityItem"

    ActivityItem:
      type: object
      required:
        - type
        - description
        - caseName
        - date
      properties:
        type:
          type: string
        description:
          type: string
        caseName:
          type: string
        date:
          type: string

    Case:
      type: object
      required:
        - id
        - name
        - client
        - status
        - daysActive
        - createdAt
      properties:
        id:
          type: integer
        name:
          type: string
        client:
          type: string
        status:
          type: string
          enum: [Active, Pending, Closed]
        description:
          type: ["string", "null"]
        daysActive:
          type: integer
        createdAt:
          type: string

    CaseInput:
      type: object
      required:
        - name
        - client
        - status
      properties:
        name:
          type: string
          minLength: 1
        client:
          type: string
          minLength: 1
        status:
          type: string
          enum: [Active, Pending, Closed]
        description:
          type: string

    CaseUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        client:
          type: string
          minLength: 1
        status:
          type: string
          enum: [Active, Pending, Closed]
        description:
          type: string

    Document:
      type: object
      required:
        - id
        - name
        - caseId
        - date
        - type
        - createdAt
      properties:
        id:
          type: integer
        name:
          type: string
        caseId:
          type: integer
        date:
          type: string
        type:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string

    DocumentInput:
      type: object
      required:
        - name
        - caseId
        - date
        - type
      properties:
        name:
          type: string
          minLength: 1
        caseId:
          type: integer
        date:
          type: string
        type:
          type: string
        notes:
          type: string

    DocumentUpdate:
      type: object
      properties:
        name:
          type: string
        caseId:
          type: integer
        date:
          type: string
        type:
          type: string
        notes:
          type: string

    Deadline:
      type: object
      required:
        - id
        - title
        - date
        - caseId
        - priority
        - createdAt
      properties:
        id:
          type: integer
        title:
          type: string
        date:
          type: string
        caseId:
          type: integer
        priority:
          type: string
          enum: [High, Medium, Low]
        createdAt:
          type: string

    DeadlineInput:
      type: object
      required:
        - title
        - date
        - caseId
        - priority
      properties:
        title:
          type: string
          minLength: 1
        date:
          type: string
        caseId:
          type: integer
        priority:
          type: string
          enum: [High, Medium, Low]

    DeadlineUpdate:
      type: object
      properties:
        title:
          type: string
        date:
          type: string
        caseId:
          type: integer
        priority:
          type: string
          enum: [High, Medium, Low]

    Task:
      type: object
      required:
        - id
        - title
        - caseId
        - assignee
        - status
        - dueDate
        - createdAt
      properties:
        id:
          type: integer
        title:
          type: string
        caseId:
          type: integer
        assignee:
          type: string
        status:
          type: string
          enum: [Open, "In Progress", Completed]
        dueDate:
          type: string
        createdAt:
          type: string

    TaskInput:
      type: object
      required:
        - title
        - caseId
        - assignee
        - dueDate
      properties:
        title:
          type: string
          minLength: 1
        caseId:
          type: integer
        assignee:
          type: string
        status:
          type: string
          enum: [Open, "In Progress", Completed]
        dueDate:
          type: string

    TaskUpdate:
      type: object
      properties:
        title:
          type: string
        caseId:
          type: integer
        assignee:
          type: string
        status:
          type: string
          enum: [Open, "In Progress", Completed]
        dueDate:
          type: string

    Interaction:
      type: object
      required:
        - id
        - caseId
        - client
        - date
        - type
        - notes
        - createdAt
      properties:
        id:
          type: integer
        caseId:
          type: integer
        client:
          type: string
        date:
          type: string
        type:
          type: string
          enum: [Phone Call, Email, Meeting, Letter]
        notes:
          type: string
        createdAt:
          type: string

    InteractionInput:
      type: object
      required:
        - caseId
        - client
        - date
        - type
        - notes
      properties:
        caseId:
          type: integer
        client:
          type: string
        date:
          type: string
        type:
          type: string
          enum: [Phone Call, Email, Meeting, Letter]
        notes:
          type: string

    InteractionUpdate:
      type: object
      properties:
        caseId:
          type: integer
        client:
          type: string
        date:
          type: string
        type:
          type: string
          enum: [Phone Call, Email, Meeting, Letter]
        notes:
          type: string