# lib/api-zod/src/generated/types/activityItem.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Case Management API
 * OpenAPI spec version: 0.1.0
 */

export interface ActivityItem {
  type: string;
  description: string;
  caseName: string;
  date: string;
}