# lib/api-zod/src/generated/types/case.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Case Management API
 * OpenAPI spec version: 0.1.0
 */
import type { CaseStatus } from './caseStatus';

export interface Case {
  id: number;
  name: string;
  client: string;
  status: CaseStatus;
  /** @nullable */
  description?: string | null;
  daysActive: number;
  createdAt: string;
}