# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Case Management API
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityItem } from './activityItem';

export interface DashboardSummary {
  totalCases: number;
  activeCases: number;
  pendingCases: number;
  upcomingDeadlines: number;
  highPriorityDeadlines: number;
  openTasks: number;
  inProgressTasks: number;
  totalDocuments: number;
  totalInteractions: number;
  recentActivity: ActivityItem[];
}