# lib/api-zod/src/generated/types/document.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Case Management API
 * OpenAPI spec version: 0.1.0
 */

export interface Document {
  id: number;
  name: string;
  caseId: number;
  date: string;
  type: string;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}